+++ src/actions/index.js
@@ -1,4 +1,4 @@
-export const REMOVE_BOOK = book => ({ type: 'REMOVE_BOOK', book });
+const REMOVE_BOOK = book => ({ type: 'REMOVE_BOOK', book });
 const ADD_BOOK = book => ({ type: 'ADD_BOOK', book });
 
-export default ADD_BOOK;
+export { ADD_BOOK, REMOVE_BOOK };
