+++ src/containers/BooksList.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import Book from './Book';
+import Book from '../components/Book';
 
 const mapStateToProps = state => state;
